third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-js/timeline-microtasks.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/timeline-test.js"></script>
<script>

var scriptUrl = "timeline-network-resource.js";

function performActions()
{
    function promiseResolved()
    {
        setTimeout(() => {}, 0);
    }

    return new Promise(fulfill => {
        var xhr = new XMLHttpRequest();
        xhr.onreadystatechange = () => xhr.readyState === 4 ? fulfill() : 0;
        xhr.onerror = fulfill;
        xhr.open("GET", "../resources/test.webp", true);
        xhr.send();
    }).then(promiseResolved);
}

function test() {
  PerformanceTestRunner.invokeAsyncWithTimeline('performActions', finish);

  function finish() {
    var event = PerformanceTestRunner.timelineModel().mainThreadEvents().find(
        e => e.name === TimelineModel.TimelineModel.RecordType.RunMicrotasks);
    PerformanceTestRunner.printTraceEventProperties(event);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Checks the RunMicrotasks event is emitted.
</p>
</body>
</html>
