src/third_party/blink/web_tests/http/tests/devtools/tracing/timeline-js/timeline-microtasks-expected.txt

Checks the RunMicrotasks event is emitted.

RunMicrotasks Properties:
{
    data : {
        microtask_count : 1
    }
    endTime : <number>
    startTime : <number>
    type : "RunMicrotasks"
}
